OBPIH-6894 Move to counting step and add redirect
