+++ src/js/hooks/cycleCount/useToCountTab.jsx
@@ -3,13 +3,14 @@ import React, { useMemo } from 'react';
 import { createColumnHelper } from '@tanstack/react-table';
 import _ from 'lodash';
 import { useSelector } from 'react-redux';
+import { useHistory } from 'react-router-dom';
 
 import cycleCountApi from 'api/services/CycleCountApi';
 import { CYCLE_COUNT_CANDIDATES } from 'api/urls';
 import { TableCell } from 'components/DataTable';
 import TableHeaderCell from 'components/DataTable/TableHeaderCell';
 import Checkbox from 'components/form-elements/v2/Checkbox';
-import { INVENTORY_ITEM_URL } from 'consts/applicationUrls';
+import { CYCLE_COUNT, INVENTORY_ITEM_URL } from 'consts/applicationUrls';
 import CycleCountStatus from 'consts/cycleCountStatus';
 import useSpinner from 'hooks/useSpinner';
 import useTableCheckboxes from 'hooks/useTableCheckboxes';
@@ -29,6 +30,7 @@ const useToCountTab = ({
   const columnHelper = createColumnHelper();
   const translate = useTranslate();
   const spinner = useSpinner();
+  const history = useHistory();
 
   const { currentLocale, currentLocation } = useSelector((state) => ({
     currentLocale: state.session.activeLanguage,
@@ -300,6 +302,7 @@ const useToCountTab = ({
     try {
       // Mocked request for start counting
       await cycleCountApi.startCount(payload, currentLocation?.id);
+      history.push(CYCLE_COUNT.countStep());
     } finally {
       spinner.hide();
     }
